+++ DynamicPriceService/ViewModels/PriceRuleViewModel.cs
@@ -3,7 +3,7 @@
 public class PriceRuleViewModel
 {
 	public int PriceRuleId { get; set; }
-	public int Increase { get; set; }
-	public int Reduction { get; set; }
+	public double Increase { get; set; }
+	public double Reduction { get; set; }
 	public TimeSpan? NoSellTime { get; set; }
 }
